+++ .github/workflows/ci.yml
@@ -10,30 +10,31 @@ env:
   PYTHONPATH: './src:./test'
 
 jobs:
-  # build:
-  #   strategy:
-  #     fail-fast: false
-  #     matrix:
-  #       os: ['ubuntu-latest']
-  #       python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
-  #   runs-on: ${{ matrix.os }}
-  #   steps:
-  #     - uses: actions/checkout@v4
+  build:
+    strategy:
+      fail-fast: false
+      matrix:
+        os: ['ubuntu-latest']
+        python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v4
 
-  #     - uses: actions/setup-python@v5
-  #       with:
-  #         python-version: ${{ matrix.python-version }}
-  #         cache: 'pip'
+      - uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+          cache: 'pip'
 
-  #     - name: setup tox
-  #       run: pip install tox
+      - name: setup tox
+        run: pip install tox
 
-  #     - name: test
-  #       run: |
-  #         TOXENV="py$(echo ${{ matrix.python-version }} | sed 's/\.//')"
-  #         tox -e "${TOXENV}"
-          
+      - name: test
+        run: |
+          TOXENV="py$(echo ${{ matrix.python-version }} | sed 's/\.//')"
+          tox -e "${TOXENV}"
+
   docs:
+    needs: build
     runs-on: ubuntu-latest
     permissions:
       contents: write
@@ -51,7 +52,7 @@ jobs:
         run: sphinx-build docs docs/_build
 
       - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3
+        uses: peaceiris/actions-gh-pages@aa83d0c2cfc3d813560e13068d3152aa21490171
         # if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
