--- docs/cli.rst
@@ -1,15 +1,0 @@
-Command Line Interface
-======================
-   
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   cli_check
-   cli_topo
-
-.. argparse::
-   :module: cycl.parser
-   :func: create_parser
-   :prog: cycl
-   :group: positional arguments
