--- docs/cli_check.rst
@@ -1,7 +1,0 @@
-cycl check
-======================
-
-.. argparse::
-   :module: cycl.parser
-   :func: create_parser
-   :path: check
