--- docs/cli_topo.rst
@@ -1,8 +1,0 @@
-cycl topo
-======================
-
-.. argparse::
-   :module: cycl.parser
-   :func: create_parser
-   :prog: cycl
-   :path: topo
