+++ docs/conf.py
@@ -1,32 +1,30 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# For the full list of built-in configuration values, see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
+import sys
+# import os
+from unittest import mock
 
-# -- Project information -----------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+MOCK_MODULES = ['boto3', 'networkx']
+for mod_name in MOCK_MODULES:
+    sys.modules[mod_name] = mock.Mock()
+
 
 project = 'cycl'
 copyright = '2025, tcm5343'
 author = 'tcm5343'
 release = '0.2.1'
 
-# -- General configuration ---------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
-
 extensions = [
-    'sphinx.ext.autodoc', 
-    'sphinxarg.ext',
-    "sphinx.ext.githubpages",
     'sphinx_rtd_theme',
+    'sphinx.ext.autodoc',
+    'sphinxarg.ext',
+    'sphinx.ext.githubpages',
+    'myst_parser',
 ]
 
 templates_path = ['_templates']
 exclude_patterns = []
 
-
-# -- Options for HTML output -------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
-
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
+
+# sys.path.insert(0, os.path.abspath('../..'))
