--- docs/index.rst
@@ -1,18 +1,0 @@
-.. cycl documentation master file, created by
-   sphinx-quickstart on Sun Feb 16 23:31:33 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-cycl documentation
-==================
-
-Add your content using ``reStructuredText`` syntax. See the
-`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
-documentation for details.
-
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   cli
