+++ requirements-docs.txt
@@ -1,6 +1,5 @@
+myst-parser==4.0.1
+sphinx_rtd_theme==3.0.2
+sphinx-argparse==0.5.2
+sphinx-autobuild==2024.10.3 
 sphinx==8.1.3
-myst-parser
-sphinx-autobuild
-sphinx-argparse
-sphinx_rtd_theme
-sphinx_rtd_theme
